Added comments and moved stuff around
